site/requirements.txt
gitdb==4.0.10
GitPython==3.1.30
Jinja2==3.1.2
MarkupSafe==2.1.1
smmap==5.0.0
